+++ pyproject.toml
@@ -4,10 +4,6 @@ version = "0.1.0"
 description = "A MCP server to upload and convert PDF files to Markdown"
 requires-python = ">=3.10"
 dependencies = [
-    "pdfplumber>=0.11.4",
-    "ocrmypdf>=16.5.0",
-    "fastmcp>=2.0.0",
-    "uvicorn>=0.30.6",
     "langchain-mcp-adapters>=0.1.0",
     "anthropic>=0.34.2",
     "httpx>=0.27.0",
@@ -17,7 +13,6 @@ dependencies = [
     "langgraph>=0.2.55",
     "anyio>=4.0.0",
     "httpcore>=1.0.0",
-    "mcp>=0.1.43",
     "fastapi>=0.115.0",
     "requests>=2.31.0",
     "langchain>=0.3.26",
