+++ src/client/client_upload.py
@@ -45,7 +45,7 @@ async def upload_file(file_path, delete_after=False):
         with open(file_path, "rb") as f:
             files = {"file": (file_name, f, "application/pdf")}
             data = {"delete_after": str(delete_after).lower()}
-            response = requests.post("http://127.0.0.1:8000/mcp/upload_pdf_tool", files=files, data=data)
+            response = requests.post("http://127.0.0.1:8001/upload/mcp/upload_pdf_tool", files=files, data=data)
             response.raise_for_status()
             logger.info(f"File {file_name} uploaded successfully")
             return file_name
